# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Federated Learning API",
        "description": "API documentation for the Medical AI Federated Learning platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/payment/webhook": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Handle Chargily Pay Webhook",
                "description": "Public endpoint for Chargily to push payment status updates. Secured by HMAC signature.",
                "operationId": "b2db396b55559250490e57e95db3efe8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook received and processed"
                    },
                    "400": {
                        "description": "Missing signature or malformed payload"
                    },
                    "403": {
                        "description": "Invalid HMAC signature"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "HttpOnly cookie issued by Sanctum after a successful `/api/verify-otp` call.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}